requirements.txt
openpyxl>=3.1.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pymilvus>=2.4.0
fastapi==0.128.5
uvicorn==0.40.0
langgraph==1.0.8
pydantic==2.12.5
redis==7.1.0
pymilvus==2.6.8
psycopg2-binary==2.9.11
openai==2.17.0
sentence-transformers==2.7.0
torch==2.10.0
python-dotenv==1.2.1